build: unset LZMADictionarySize high value.
With compression level "ultra64" for LZMA, default value is 65536, which
is far enough (the installer file was only 100kib bigger, it doesn't
make that big of a difference in the end).
With the original value, the CI build would fail with "Out of memory"
error.

See also: https://jrsoftware.org/ishelp/index.php?topic=setup_lzmadictionarysize